requirements.txt
timm
absl-py==1.4.0
cachetools==5.3.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
markdown==3.4.4
oauthlib==3.2.2
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
tensorboard==2.14.0
tensorboard-data-server==0.7.1
torch-summary==1.4.5
urllib3==1.26.16
werkzeug==2.3.7
pandas==0.25.3
tables==3.7.0
matplotlib==3.1.3
pywavelets==1.4.1
numpy==1.19.5
pyyaml
ipython
scikit-learn
pytorch_wavelets
loralib